feat: util
