+++ src/shared/state/stores/UtilityStore.ts
@@ -1,6 +1,6 @@
 import path from "path";
 import { AbstractStore } from "./AbstractStore";
-import { getSubFiles, getSubFolders } from "../../util";
+import { getSubFiles, getSubFolders } from "./util";
 import { ReferenceFile } from "../items/ReferenceFile";
 import { Time } from "../../../shared/time";
